backend/requirements.txt
firebase-admin
Flask
pyautogui
pyperclip
python-dotenv
